+++ src/TypeInspectors/GetStrictTypes.php
@@ -62,7 +62,7 @@ class GetStrictTypes
      * @return array
      *         the list of type(s) that this item can be
      */
-    public function __invoke($item)
+    public function getStrictTypes($item)
     {
         return self::from($item);
     }